requirements.txt
 chromadb
  sentence-transformers
boto3
 google-cloud-storage
sqlalchemy
psycopg